requirements.txt
gensim
flask
jieba
pyecharts
pygal
charts
selenium
js2xml
parsel
MyQR
Pillow